Started the third optimized implementation.
Introduced the third implementation for MV multiplication. This one has
the same structure as plainKernel, but somewhat different naming.

This version groups the dot products in tiles, so that the vector is
efficiently cached for all block threads. Runtimes are slow but this
should get fixed soon. Code does not handle the case where columns are
not evenly divided by the tile size.

Moved the sizeInfo typedef to common.h, since this will be used from
multiple places.