Modified 'tile' concept for optimized multiplication.
Added missing common.h file.

Added atomicAdd function for double types (taken from nVidia website).

Grid is now 2D, and each block assigns its threads to load a chunk of
the multiplied vector to shared memory. This chunk is used for the dot
product, and the results are reduced with the atomicAdd function.

Changed matrix storage type to column major. Essentialy the only thing
changed is the load order in the kernel, so that loads are coalesced.

Grid and block sizes are defined according to tile size and block
threads.