Changed grid dimensions, which fixed out of bounds read bug for optimized kernel.
Added comment to calculation of tile origin points, to avoid confusion.
Grid's x dimension maps to the columns, and y dimension to the rows.

Grid dimensions were the opposite of what they should be; also, reduced
size of grid by 1 when the input matrix was exactly the dimensions of
the block/tile size. This is done by dividing with the required
tile/block size + 1.

Speedup of 5x!!!! :)