Added correctness check for plain kernel.
If the runCPU is set to true, plainKernel will run the CPU version of MV
and subsequently compare the results with the CUDA implementation. Shows
a helpful message when an error is encountered, otherwise an indication
that the test was passed successfully. Doubles are compared via the
epsilon method, i.e. comparing their (absolute) difference to a really
small number.