Fixed optimized kernel grid size and added correctness test.
Modified CPU matrix multiplication for the optimized kernel, to use the
matrix in column-major form.

Added a correctness tests with messages for failing or passing the test;
this compares the CUDA results with the CPU results if runCPU is true.
EPSILON is a constant defined before main, so that the difference
between results is not less than that.

Fixed grid size for the optimized kernel. Previously, the calculation
was missing the last elements for an awkwardly sized matrix (like
1025x1025). Now, round up integer division is used (so 1024 columns use
2 tiles, 1025 columns use 3 tiles).

Enabled unit tests for plain and optimized kernel in main.