Fixed CUBLAS array sizes.
Since CUBLAS expects a column-major matrix, and we pass a row-major one,
cublasDgemv needs A transposed (as it already is done). However, the
dimensions should be inverted, because the host essentially transfers
the transposed A matrix, which later is transposed again to produce
correct results. Basically this commit changes rows <=> cols and the
leading dimensions for cublasSetMatrix and cublasDgemv.